fix compilation warning reported by automated multi-platform scan
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@34794 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/algo/structure/cd_utils/cuAlign.cpp
@@ -971,7 +971,7 @@ string ddAlignInfo(TDendiag * pGuideDD)
                 CRef<CSeq_id> GuideMasterSeqID=*(pid);
                 CRef<CSeq_id> GuideSeqID=*(++pid);
                 
-                sprintf(buf,"[%s]/[%s](%d)  ",GetSeqIDStr(GuideMasterSeqID).c_str(),GetSeqIDStr(GuideSeqID).c_str(),pGuideDD->size());
+                sprintf(buf,"[%s]/[%s](%d)  ",GetSeqIDStr(GuideMasterSeqID).c_str(),GetSeqIDStr(GuideSeqID).c_str(),(int)pGuideDD->size());
                 if(!iDst){
                     ret+=buf;
                 }
